requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
websockets==12.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
deepface==0.0.93
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.4
python-dotenv==1.0.1
fpdf2==2.7.9
matplotlib==3.8.4
groq==0.9.0 
httpx==0.27.0
ultralytics 